+++ src/Data/IptcInterface.php
@@ -11,6 +11,11 @@
 
 namespace PHPExif\Common\Data;
 
+use PHPExif\Common\Data\ValueObject\Caption;
+use PHPExif\Common\Data\ValueObject\Copyright;
+use PHPExif\Common\Data\ValueObject\Credit;
+use PHPExif\Common\Data\ValueObject\Headline;
+
 /**
  * IptcInterface
  *
@@ -21,4 +26,67 @@ namespace PHPExif\Common\Data;
  */
 interface IptcInterface
 {
+    /**
+     * Accessor for the headline
+     *
+     * @return Headline
+     */
+    public function getHeadline();
+
+    /**
+     * Returns new instance with updated headline
+     *
+     * @param Headline $headline
+     *
+     * @return ExifInterface
+     */
+    public function withHeadline(Headline $headline);
+
+    /**
+     * Accessor for the credit
+     *
+     * @return Credit
+     */
+    public function getCredit();
+
+    /**
+     * Returns new instance with updated credit
+     *
+     * @param Credit $credit
+     *
+     * @return ExifInterface
+     */
+    public function withCredit(Credit $credit);
+
+    /**
+     * Accessor for the copyright
+     *
+     * @return Copyright
+     */
+    public function getCopyright();
+
+    /**
+     * Returns new instance with updated copyright
+     *
+     * @param Copyright $copyright
+     *
+     * @return ExifInterface
+     */
+    public function withCopyright(Copyright $copyright);
+
+    /**
+     * Accessor for the caption
+     *
+     * @return Caption
+     */
+    public function getCaption();
+
+    /**
+     * Returns new instance with updated caption
+     *
+     * @param Caption $caption
+     *
+     * @return ExifInterface
+     */
+    public function withCaption(Caption $caption);
 }
